Make Register a shared class, to help with sequence generation
