+++ generator/CMakeLists.txt
@@ -1,5 +1,4 @@
 cmake_minimum_required(VERSION 3.14)
-project(generator)
 
 # Set the C++ standard
 set(CMAKE_CXX_STANDARD 17)
@@ -16,6 +15,7 @@ endif()
 
 # Set the include directories
 include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include)
+include_directories(${PROJECT_SOURCE_DIR}/shared)
 
 # Add the source files
 file(GLOB SOURCES "src/*.cpp")
