+++ mock/CMakeLists.txt
@@ -1,5 +1,4 @@
 cmake_minimum_required(VERSION 3.14)
-project(mock)
 
 # Set the C++ standard
 set(CMAKE_CXX_STANDARD 17)
@@ -8,6 +7,7 @@ set(CMAKE_CXX_STANDARD_REQUIRED ON)
 # Set the include directories
 include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include)
 include_directories(${PROJECT_SOURCE_DIR}/IPbus/inc)
+include_directories(${PROJECT_SOURCE_DIR}/shared/include)
 
 # Add the source files
 file(GLOB SOURCES "src/*.cpp")
